\name{blaunetgui}
\alias{blaunetgui}
\title{
Run Blau status analysis in graphic user interface (GUI).
}
\description{
Load graphic user interface (GUI) for Blaunet package.
}
\usage{
blaunetgui()
}
\details{
blaunetgui has the following five main capabilities, it can: 

1) identify a list of possible salient dimensions; 

2) calculate, plot, and analyze niches for social entities by measuring the social distance along the salient dimensions between individuals affiliated with them; 

3) capture niche dynamics cross-sectionally by calculating the intensity of exploitation from the carrying capacity and the membership rate; 

4) analyze the niche movement longitudinally by estimating the predicted niche movement equations; and 

5) generate Blau bubbles for individuals, thereby allowing the study of interpersonal influence of similar others even with no or limited network information.
}
\value{No return value}
\examples{
\donttest{
blaunetgui()
}
}
