\name{gss74_87}
\alias{gss74_87}
\docType{data}
\title{
Example attribute dataset from General Social Survey (GSS).
}
\description{
The attribute dataset contains 13,865 respondents surveyed between 1974 and 1987 and their attribute and group affiliation information. The attribute data includes education level and occupational prestige for constructing Blau space, along with group affiliation data in 16 types of organizations
}
\usage{data(gss74_87)}
\source{
This dataset isexcerpted from the General Social Survey, 1972-2012, available from both the NORC website(http://gss.norc.org/get-the-data/) and the Interuniversity Consortium for Political and Social Research (ICPSR) website (http://www.icpsr.umich.edu/icpsrweb/NACJD/studies/34802).  
}
\value{
Returns a data frame containing the gss74_87 attribute file.
}
\examples{
data(gss74_87)
}
\keyword{datasets}
