\name{schlattr}
\alias{schlattr}
\docType{data}
\title{
Example attribute dataset from Add Health.
}
\description{
The attribute dataset contains 100 respondents with 20 individual characteristics and group affiliation data in 5 kinds of organizations. 
}
\usage{data(schlattr)}
\source{
The attribute data of 100 respondents are randomly drawn from the wave I public use data of National Longitudinal Study of Adolescent to Adult Health (Add Health), available from the Interuniversity Consortium for Political and Social Research (ICPSR) website, data number 21600.
}
\value{
A data frame containing the school attribute file.
}
\examples{
data(schlattr)
}
\keyword{datasets}
