\name{schlnet}
\alias{schlnet}
\docType{data}
\title{
Example network dataset simulated based on the attribute dataset from Add Health.
}
\description{
The network dataset contains an adjacent matrix of 100 individuals. 
}
\usage{data(schlnet)}
\source{
The network dataset is not real.  It is simulated with an Exponential family random graph (ERG) model based on the attribute dataset from Add Health
}
\value{
A dataframe containing the school network file.
}
\examples{
data(schlnet)
}
\keyword{datasets}
