\name{blaunetgui}
\alias{blaunetgui}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Run Blau status analysis in graphic user interface (GUI).
}
\description{
Load graphic user interface (GUI) for BlauNet package.
}
\usage{
blaunetgui()
}
\details{
blaunetgui has the following five main capabilities, it can: 

1) identify a list of possible salient dimensions; 

2) calculate, plot, and analyze niches for social entities by measuring the social distance along the salient dimensions between individuals affiliated with them; 

3) capture niche dynamics cross-sectionally by calculating the intensity of exploitation from the carrying capacity and the membership rate; 

4) analyze the niche movement longitudinally by estimating the predicted niche movement equations; and 

5) generate Blau bubbles for individuals, thereby allowing the study of interpersonal influence of similar others even with no or limited network information.
}
%%\references{
%% ~put references to the literature/web site here ~
%%}
%%\author{
%%  ~~who you are~~
%%}
%%\note{
%%  ~~further notes~~
%%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%%\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
%%}
\examples{
##not running
##blaunetgui()
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%%\keyword{ ~kwd1 }
%%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
